Move PlantCard to its own folder
